+++ Changelog.md
@@ -5,6 +5,7 @@ Released on ????-??-??.
 
 #### Added
 - Added `element(withIdentifier:labels:labelComparisonOperator:)`
+- Improve launch options type safety
 
 ## [1.2.0](https://github.com/PGSSoft/AutoMate/releases/tag/1.2.0)
 Released on 2017-03-29.